take apart table creation for interlis and standart themes
